requirements.txt
Django==3.2.12
torch==2.6.0
torchvision==0.21.0
Pillow==8.4.0
